refactor(dashboard): decompose LiveTab into cards, modals, and data hook
