+++ packages/frontend/src/components/dashboard/modals/index.ts
@@ -1,0 +1,15 @@
+/**
+ * @file index.ts
+ *
+ * Barrel export for Live Metrics dashboard modals.
+ */
+
+export * from './VelocityModal';
+export * from './ProviderPulseModal';
+export * from './ModelPulseModal';
+export * from './TimelineModal';
+export * from './ModelTimelineModal';
+export * from './RequestsModal';
+export * from './ConcurrencyModal';
+export * from './StatsModal';
+export * from './LiveDashboardModal';
